% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentroidSpeices.R
\name{CentroidSpeices}
\alias{CentroidSpeices}
\title{Get the species in the centroid}
\usage{
CentroidSpeices(species, centroid)
}
\arguments{
\item{species}{a string containing all species name and country (if any) in the end, separate by ','.}

\item{centroid}{the result generated by the function Getcentroid}
}
\value{
the gbif data that inside the centroid
}
\description{
Get the species in the centroid
}
\examples{
centroid <- Getcentroid('Canada', 8)
CentroidSpecies('Pinus contorta,Danaus plexippus,Canada', centroid)
}
