% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checkdate.R
\name{Checkdate}
\alias{Checkdate}
\title{Check the completeness of the date record}
\usage{
Checkdate(species, number = 10000)
}
\arguments{
\item{species}{a string containing all species name and country (if any) in the end, separate by ','.}

\item{number}{how many record is going to be checked}
}
\value{
the gbif data with a datequality column and a table showing the rate of missing whole eventdate, month and day.
}
\description{
Check the completeness of the date record
}
\examples{
Checkdate('Pinus contorta,Danaus plexippus,Canada', number = 20000)
}
