% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Getcentroid.R
\name{Getcentroid}
\alias{Getcentroid}
\title{Generate centroid area WKT data}
\usage{
Getcentroid(country, buffer_range)
}
\arguments{
\item{country}{the name of interested country}

\item{buffer_range}{the disstance between the centroid and the border, buffer_range = 1 roughly equals to 118 km}
}
\value{
centroid area WKT data
}
\description{
Generate centroid area WKT data
}
\examples{
Getcentroid('china', 8)
}
