% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Getoccdat.R
\name{Getoccdat}
\alias{Getoccdat}
\title{Get occdat via occ command}
\usage{
Getoccdat(species)
}
\arguments{
\item{species}{a string containing all species name and country (if any) in the end, separate by ','.}
}
\value{
a list of occdat, exist species names, all species names, exist speices number, all species number, country name and the indicater of specific country
}
\description{
Get occdat via occ command
}
\examples{
data <- Getoccdat('Pinus contorta,Danaus plexippus,Canada')
}
