% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_occdat.R
\name{leaflet_occdat}
\alias{leaflet_occdat}
\title{Plot the occurrence data via map_leaftel}
\usage{
leaflet_occdat(datalist)
}
\arguments{
\item{datalist}{the result generated by the function Getoccdat}
}
\value{
the occurrence map with legends
}
\description{
Plot the occurrence data via map_leaftel
}
\examples{
name <- Readname('Pinus contorta,Danaus plexippus,Canada')
data <- Getoccdat(name)
leaflet_occdat(data)
}
