% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgcd.R
\name{vector_multiply}
\alias{vector_multiply}
\title{Parallel vector multiplication.}
\usage{
vector_multiply(x, y, gcd = FALSE, omp = FALSE)
}
\arguments{
\item{x}{A numeric vector of length n.}

\item{y}{A numeric vector of length n.}

\item{gcd}{A boolean to use gcd or not.}

\item{omp}{A boolean to use OpenMP or not, if gcd is TRUE OpenMP will not be used.}
}
\value{
The output is the product of x and y.
}
\description{
\code{vector_multiply} multiplies two vectors together using either OpenMP or
Grand Central Dispatch, provided one is available.
}
\examples{
x <- 1:3
y <- 10:12
vector_multipy(x,y)
}
\author{
Jonthan Lisic, \email{jlisic@gmail.com}
}
