% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentroidSpeices.R
\name{CentroidSpeices}
\alias{CentroidSpeices}
\title{Get the species in the centroid}
\usage{
CentroidSpeices(namelist, centroid)
}
\arguments{
\item{namelist}{the result generated by the function Readname}

\item{centroid}{the result generated by the function Getcentroid}
}
\value{
the gbif data that inside the centroid
}
\description{
Get the species in the centroid
}
\examples{
name <- Readname('Pinus contorta,Danaus plexippus,Canada')
centroid <- Getcentroid('Canada', 8)
CentroidSpecies(name, centroid)
}
