% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checkdate.R
\name{Checkdate}
\alias{Checkdate}
\title{Check the completeness of the date record}
\usage{
Checkdate(namelist, number = 10000)
}
\arguments{
\item{namelist}{the result generated by the function Readname}

\item{number}{how many record is going to be checked}
}
\value{
the gbif data with a datequality column
}
\description{
Check the completeness of the date record
}
\examples{
name <- Readname('Pinus contorta,Danaus plexippus,Canada')
Checkdate(name, number = 20000)
}
