% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Getoccdat.R
\name{Getoccdat}
\alias{Getoccdat}
\title{Get occdat via occ command}
\usage{
Getoccdat(namelist)
}
\arguments{
\item{namelist}{the result generated by the function Readname}
}
\value{
a list of occdat, exist species names, all species names, exist speices number, all species number, country name and the indicater of specific country
}
\description{
Get occdat via occ command
}
\examples{
name <- Readname('Pinus contorta,Danaus plexippus,Canada')
data <- Getoccdat(name)
}
